plastic-django/pyproject.toml
[project]
name = "plastic-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.3",
    "ruff>=0.7.3",
    "django-extensions>=3.2.3",
]
